New Val-Loader https://github.com/webpack-contrib/val-loader
